mmsci-exps/requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
beautifulsoup4==4.12.3
bert-score==0.3.13
blessed==1.19.1
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
dill==0.3.8
deepspeed==0.9.5
einops==0.8.0
eval==0.0.5
evaluate==0.4.2
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
flash-attn==2.5.8
fsspec==2024.3.1
ftfy==6.2.0
gpustat==1.0.0
huggingface-hub==0.23.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==11.495.46
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.4
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pylatexenc==2.10
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
peft==0.11.1
regex==2024.5.15
requests==2.32.2
rouge_score==0.1.2
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.5
sympy==1.12
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
transformers-stream-generator==0.0.5
triton==2.3.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
openai==1.32.0